rename segmentation prompt
